+++ src/main/java/com/oneclick/productservice/domain/factory/ProductFactory.java
@@ -2,7 +2,9 @@ package com.oneclick.productservice.domain.factory;
 
 import com.oneclick.productservice.domain.Product;
 
+import java.math.BigDecimal;
+
 public interface ProductFactory {
 
-    Product create(Long id, String name, String description, Double price);
+    Product create(Long id, String name, String description, BigDecimal price);
 }
