+++ src/main/java/com/oneclick/productservice/dto/ProductResponse.java
@@ -1,17 +1,19 @@
 package com.oneclick.productservice.dto;
 
+import java.math.BigDecimal;
+
 public class ProductResponse {
     private long id;
     private String name;
     private String description;
-    private double price;
+    private BigDecimal price;
     private String type;
 
     public ProductResponse() {
 
     }
 
-    public ProductResponse(long id, String name, String description, double price, String type) {
+    public ProductResponse(long id, String name, String description, BigDecimal price, String type) {
         this.id = id;
         this.name = name;
         this.description = description;
@@ -43,11 +45,11 @@ public class ProductResponse {
         this.description = description;
     }
 
-    public double getPrice() {
+    public BigDecimal getPrice() {
         return price;
     }
 
-    public void setPrice(double price) {
+    public void setPrice(BigDecimal price) {
         this.price = price;
     }
